feature: add base for the signal websocket client
